LayoutTests/css3/scroll-snap/scroll-snap-initial-expected.txt

Tests 'initial' on scroll snap properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS noInitial.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS noInitial.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS noInitial.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS noInitial.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-destination') is "10% 11%"
PASS noInitial.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(noInitial).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS initialType.style['-webkit-scroll-snap-type'] is "initial"
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-type') is "none"
PASS initialType.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS initialType.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS initialType.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS initialType.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(initialType).getPropertyValue('-webkit-scroll-snap-coordinate') is "15px 122px"
PASS initialXPoints.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS initialXPoints.style['-webkit-scroll-snap-points-x'] is "initial"
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-points-x') is "none"
PASS initialXPoints.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS initialXPoints.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-destination') is "50% 70%"
PASS initialXPoints.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(initialXPoints).getPropertyValue('-webkit-scroll-snap-coordinate') is "100% 50%"
PASS initialYPoints.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS initialYPoints.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS initialYPoints.style['-webkit-scroll-snap-points-y'] is "initial"
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-points-y') is "none"
PASS initialYPoints.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-destination') is "10px 20px"
PASS initialYPoints.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(initialYPoints).getPropertyValue('-webkit-scroll-snap-coordinate') is "50% 100%"
PASS initialDestination.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-type') is "proximity"
PASS initialDestination.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS initialDestination.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS initialDestination.style['-webkit-scroll-snap-destination'] is "initial"
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS initialDestination.style['-webkit-scroll-snap-coordinate'] is ""
PASS window.getComputedStyle(initialDestination).getPropertyValue('-webkit-scroll-snap-coordinate') is "10px 10px"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-type'] is ""
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-type') is "mandatory"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-points-x'] is ""
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-x') is "repeat(100%)"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-points-y'] is ""
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-points-y') is "repeat(100%)"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-destination'] is ""
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-destination') is "10% 10%"
PASS initialSnapCoordinate.style['-webkit-scroll-snap-coordinate'] is "initial"
PASS window.getComputedStyle(initialSnapCoordinate).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS allInitial.style['-webkit-scroll-snap-type'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-type') is "none"
PASS allInitial.style['-webkit-scroll-snap-points-x'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-points-x') is "none"
PASS allInitial.style['-webkit-scroll-snap-points-y'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-points-y') is "none"
PASS allInitial.style['-webkit-scroll-snap-destination'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS allInitial.style['-webkit-scroll-snap-coordinate'] is "initial"
PASS window.getComputedStyle(allInitial).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS allInitialY.style['-webkit-scroll-snap-type'] is "initial"
PASS window.getComputedStyle(allInitialY).getPropertyValue('-webkit-scroll-snap-type') is "none"
PASS allInitialY.style['-webkit-scroll-snap-points-x'] is "initial"
PASS window.getComputedStyle(allInitialY).getPropertyValue('-webkit-scroll-snap-points-x') is "none"
PASS allInitialY.style['-webkit-scroll-snap-points-y'] is "initial"
PASS window.getComputedStyle(allInitialY).getPropertyValue('-webkit-scroll-snap-points-y') is "none"
PASS allInitialY.style['-webkit-scroll-snap-destination'] is "initial"
PASS window.getComputedStyle(allInitialY).getPropertyValue('-webkit-scroll-snap-destination') is "0px 0px"
PASS allInitialY.style['-webkit-scroll-snap-coordinate'] is "initial"
PASS window.getComputedStyle(allInitialY).getPropertyValue('-webkit-scroll-snap-coordinate') is "none"
PASS successfullyParsed is true

TEST COMPLETE
